refactor: cardListProvider 분리
