+++ src/App.tsx
@@ -1,23 +1,15 @@
-import { useState } from 'react';
 import { Outlet } from 'react-router-dom';
-import { Card, CardContext } from './context/CardContext';
-
-import { getLocalStorage } from './utils/localStorage';
+import { CardListProvider } from './context/CardContext';
 import GlobalStyle from './styles/global';
 import Header from './components/Header';
-import { LOCAL_STORAGE_KEY } from './constants';
 
 function App() {
-  const [cardList, setCardList] = useState(
-    getLocalStorage<Card[]>(LOCAL_STORAGE_KEY.CARD_LIST, []),
-  );
-
   return (
-    <CardContext.Provider value={{ cardList, setCardList }}>
+    <CardListProvider>
       <GlobalStyle />
       <Header />
       <Outlet />
-    </CardContext.Provider>
+    </CardListProvider>
   );
 }
