+++ src/types/index.ts
@@ -1,0 +1,10 @@
+type Card = {
+  id: string;
+  cardNumbers: string[];
+  cardExpirationDate: string[];
+  cardOwner: string[];
+  cardCVC: string[];
+  cardPWD: string[];
+};
+
+export type { Card };
